// src/app/models/DatosBancariosSearchAnexosItem.ts
export class DatosBancariosSearchAnexosItem {
  idPersona: String;
  idInstitucion: String;
  idCuenta: String;
  idMandato: String;
  idAnexo: String;
  descripcion: String;
  tipoMandato: String;
  tipo: String;
  fechaUso: String;
  firmaFecha: String;
  firmaLugar: String;

  constructor() {}
}